graph working and styled
