+++ app/src/main/res/values-en/strings.xml
@@ -37,5 +37,6 @@
     <string name="airpressure">Air Pressure</string>
     <string name="airpressign">hPa</string>
 
-
+    <string name="graphdeschum">Humidity</string>
+    <string name="graphdesctemp">Temperature</string>
 </resources>
